requirements.txt
discord.py
yt-dlp
google-generativeai
edge-tts
psutil
python-dotenv
ffmpeg-python
PyNaCl
Pillow
spotipy
